added checks for jwt verification
